Made gradle run work
