+++ app/client/build.gradle.kts
@@ -4,12 +4,15 @@ plugins {
 }
 
 dependencies {
-    // Shared dependencies
+    // Shared Flaggi library
     implementation(project(":shared"))
 
-    // Use JUnit Jupiter for testing.
+    // JUnit for testing
     testImplementation(libs.junit.jupiter)
     testRuntimeOnly("org.junit.platform:junit-platform-launcher")
+
+    // Application dependencies
+    // implementation("com.google.guava:guava:31.1-jre")
 
     // JSON file manipulation
     implementation("com.fasterxml.jackson.core:jackson-databind:2.17.1")
@@ -18,21 +21,15 @@ dependencies {
 
     // JSON object for Github release fetching
     implementation("org.json:json:20230618")
-
-    // Application dependencies
-    implementation(libs.guava)
 }
 
 application {
-    mainClass.set("flaggi.App")
-}
-
-tasks.build {
-    dependsOn(tasks.shadowJar)
+    mainClass.set("flaggiclient.App")
 }
 
 tasks.withType<com.github.jengelman.gradle.plugins.shadow.tasks.ShadowJar> {
     enabled = true
+    from(project(":shared").sourceSets["main"].output)
     archiveBaseName.set("Flaggi-client")
     archiveVersion.set("1.0.0")
     archiveClassifier.set("") // Removes the "-all" suffix
@@ -42,4 +39,16 @@ tasks.withType<com.github.jengelman.gradle.plugins.shadow.tasks.ShadowJar> {
     doLast {
         println("Client Shadow JAR has been created at: ${archiveFile.get().asFile.absolutePath}")
     }
+}
+
+tasks.named<JavaExec>("run") {
+    dependsOn(tasks.named<com.github.jengelman.gradle.plugins.shadow.tasks.ShadowJar>("shadowJar"))
+
+    classpath = sourceSets["main"].runtimeClasspath + files(tasks.named<com.github.jengelman.gradle.plugins.shadow.tasks.ShadowJar>("shadowJar").get().archiveFile)
+
+    mainClass.set(application.mainClass)
+}
+
+tasks.build {
+    dependsOn(tasks.shadowJar)
 }
