--- app/client/src/main/java/flaggi/constants/WidgetTags.java
@@ -1,44 +1,0 @@
-/*
- * Author: Matěj Šťastný
- * Date created: 11/6/2024
- * Github link: https://github.com/kireiiiiiiii/Flaggi
- *
- *
- *
- *
- *
- *
- *
- *
- *
- *
- *
- *
- *
- *
- *
- *
- *
- *
- *
- *
- */
-
-package flaggi.constants;
-
-/**
- * Constant class with widget tag data.
- * 
- */
-public class WidgetTags {
-
-    public static final String MENU_ELEMENTS = "menu";
-    public static final String GAME_ELEMENTS = "game";
-    public static final String ENEMY_PLAYER = "enemy";
-    public static final String PAUSE_MENU = "pause";
-    public static final String LOBBY = "lobby";
-
-    public static final String ENVIRONMENT = "environment";
-    public static final String PROJECTILES = "projectiles";
-
-}
