+++ app/client/src/main/java/flaggiclient/constants/WidgetTags.java
@@ -1,0 +1,44 @@
+/*
+ * Author: Matěj Šťastný
+ * Date created: 11/6/2024
+ * Github link: https://github.com/kireiiiiiiii/Flaggi
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ */
+
+package flaggiclient.constants;
+
+/**
+ * Constant class with widget tag data.
+ *
+ */
+public class WidgetTags {
+
+    public static final String MENU_ELEMENTS = "menu";
+    public static final String GAME_ELEMENTS = "game";
+    public static final String ENEMY_PLAYER = "enemy";
+    public static final String PAUSE_MENU = "pause";
+    public static final String LOBBY = "lobby";
+
+    public static final String ENVIRONMENT = "environment";
+    public static final String PROJECTILES = "projectiles";
+
+}
