+++ app/client/src/main/java/flaggiclient/constants/ZIndex.java
@@ -1,0 +1,53 @@
+/*
+ * Author: Matěj Šťastný
+ * Date created: 11/6/2024
+ * Github link: https://github.com/kireiiiiiiii/Flaggi
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ */
+
+package flaggiclient.constants;
+
+/**
+ * Constant class with z-layer data.
+ *
+ */
+public class ZIndex {
+
+    // GLOBAL
+    public static final int BACKGROUND = 1;
+
+    public static final int CONNECTION = 3;
+
+    // MENU
+    public static final int MENU_SCREEN = 2;
+    public static final int TOAST = 30;
+    public static final int PAUSE_SCREEN = 20;
+
+    // GAME
+    public static final int FLOOR = 2;
+    public static final int PLAYER = 11;
+    public static final int OTHER_PLAYERS = 10;
+    public static final int ENVIRONMENT_TOP = 12;
+    public static final int ENVIRONMENT_BOTTOM = 9;
+    public static final int HUD = 18;
+
+}
