+++ app/server/build.gradle.kts
@@ -4,7 +4,7 @@ plugins {
 }
 
 dependencies {
-    // Shared dependencies
+    // Shared Flaggi library
     implementation(project(":shared"))
 
     // JUnit for testing
@@ -12,7 +12,9 @@ dependencies {
     testRuntimeOnly("org.junit.platform:junit-platform-launcher")
 
     // Application dependencies
-    implementation("com.google.guava:guava:31.1-jre")
+    // implementation("com.google.guava:guava:31.1-jre")
+
+    // JSON file manipulation
     implementation("com.fasterxml.jackson.core:jackson-databind:2.15.0")
 }
 
@@ -24,8 +26,17 @@ tasks.build {
     dependsOn(tasks.shadowJar)
 }
 
+tasks.named<JavaExec>("run") {
+    dependsOn(tasks.named<com.github.jengelman.gradle.plugins.shadow.tasks.ShadowJar>("shadowJar"))
+
+    classpath = sourceSets["main"].runtimeClasspath + files(tasks.named<com.github.jengelman.gradle.plugins.shadow.tasks.ShadowJar>("shadowJar").get().archiveFile)
+
+    mainClass.set(application.mainClass)
+}
+
 tasks.withType<com.github.jengelman.gradle.plugins.shadow.tasks.ShadowJar> {
     enabled = true
+    from(project(":shared").sourceSets["main"].output)
     archiveBaseName.set("Flaggi-server")
     archiveVersion.set("1.0.0")
     archiveClassifier.set("") // Removes the "-all" suffix
